Backend/MyPortfolio/requirements.txt
django
djangorestframework
environs
mysqlclient
whitenoise
djangorestframework-simplejwt
django-cors-headers
